requirements.txt
amzqr
gradio==4.19.2
Pillow
numpy